pyproject.toml
[project]
name = "ai-cr"
version = "0.1.0"
description = "CR ai agent"
readme = "README.md"
authors = [
  {name = "Raven", email = "152022920+Raven-002@users.noreply.github.com"}
]
maintainers = [
  {name = "Raven", email = "152022920+Raven-002@users.noreply.github.com"}
]
# https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 1 - Planning",
  "Environment :: Console",

  "Intended Audience :: Developers",
  "Topic :: Software Development :: Quality Assurance",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
license = "MIT"
dependencies = [
  "gitpython>=3.1.45",
  "ollama>=0.6.1",
  "pyyaml>=6.0.3",
  "typer",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "mypy", # Another type checker
    "basedpyright", # Another type checker/linter
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/Raven-002/ai_cr/issues"
changelog = "https://github.com/Raven-002/ai_cr/blob/master/changelog.md"
homepage = "https://github.com/Raven-002/ai_cr"

[project.scripts]
ai_cr = "ai_cr.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.mypy]
strict = true
exclude = "build|dist|.venv.*|__pycache__"
disable_error_code = [
  "import-untyped"
]

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = [
  "no-untyped-def"
]

[tool.basedpyright]
# Strict type checking
typeCheckingMode = "strict"
reportMissingTypeStubs=false
# Exclude build/dist/venv artifacts
exclude = [
    "build",
    "dist",
    ".venv*",
    "**/__pycache__"
]

[tool.uv]
package = true
